requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
cssbeautifier==1.15.4
djlint==1.36.4
EditorConfig==0.17.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsbeautifier==1.15.4
json5==0.12.1
MarkupSafe==3.0.2
packaging==25.0
pathspec==0.12.1
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
six==1.17.0
tqdm==4.67.1
waitress==3.0.2
Werkzeug==3.1.3
